Map drm detected timing to hdmi tx timing table
This patch will try to find a video mode-id, from tx timing table, using drm
detected mode parameteres. If match is found uses timing parameters in mode
table to program tx. If match is not found, mode-id is tagged as CUSTOM and
drm timing is used to program TX IP.

Signed-off-by: Rohit Consul <rohitco@xilinx.com>