DP159: Write register 0x0A last, it has commit bit.
The datasheet is not clear on how the commit works. It says the
settings are not applied immediately, but it does not explain
if a write order must be used, if 0x0A must be written last, or
when the actual values are applied. To be sure, write 0x0A last.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>