V4L: Add initial Driver support for Xilinx Video Demosaic IP(HLS)
This driver supports the Xilinx Video Demosaic IP.It will support only
the following configuration of the IP in Vivado GUI :
	- 8 bit
	- High Quality Interpolation
	- No zipper

All other configurations of this IP are not supported by the driver.
The driver has been tested with 4.6 version of the kernel and 2016.3
version of Vivado.

The following features are currently missing in the driver support
and will be added in a later release :
	- GPIO reset support enumerated through Device Tree
	The driver currently lacks support to reset the IP.